it's a fucking mess. but we have a compiling vertex shader.

+++ Engine/src/engine/render/Material.cpp
@@ -13,7 +13,7 @@ namespace se::render
             const auto& lightSetup = renderer->GetLightSetup();
             if (m_CachedLightSetup != lightSetup || !m_PlatformResourcesCreated)
             {
-                m_ShaderSettings.SetSetting("numLights", lightSetup.pointLights.size());
+                m_ShaderSettings.SetSetting("numLights", (int)lightSetup.pointLights.size());
                 DestroyPlatformResources();
                 m_CachedLightSetup = lightSetup;
             }
